Did below: Use constructor injection: In the PostDataAsyncTask class, you are using a member variable to store the Context object. This is not recommended as it can lead to memory leaks. Instead, you can pass the Context object to the constructor of the class and store it in a final member variable. This way, you can be sure that the Context object will not change throughout the life of the AsyncTask.
Add error handling: In the doInBackground method, you are catching IOException and returning null if an exception occurs. This is not ideal as it does not provide any information about the error that occurred. Instead, you can catch the exception and return an error message or code that can be used to identify the error.

Use constants for URLs and MIME types: In the doInBackground method, you are hardcoding the URL and the MIME type. This is not recommended as it can lead to errors if the URL or the MIME type changes. Instead, you can define constants for the URL and the MIME type and use them in the code.